--- README.md
@@ -1,16 +1,0 @@
-# POKEDEX
-
-Nome dos integrantes: 
-- Alisson Araujo Marques Moreira
-- Franciane Brito
-- José Victor Tiburtino Ferreira
-
-Link do Surge: www.link.com.br
-
-O que funciona:
-- Item 1
-- Item 2
-
-O que não funciona: 
-- Item 3
-- Item 4
