+++ projeto-pokedex/src/Components/Buttons/styled.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 export const Button = styled.button`
-  border: none;
+  border: 1px solid red;;
   outline: none !important;
   cursor: pointer;
   padding: 8px;
