+++ projeto-pokedex/src/Screens/HomePage.js
@@ -1,6 +1,5 @@
 import React, { useContext, useEffect } from "react";
 import { CardMain } from '../Components/CardPokemon/styled'
-import Header from "../Components/Header/Header";
 import CardPokemon from "../Components/CardPokemon/CardPokemon";
 import GlobalStateContext from "../Global/GlobalStateContext";
 
@@ -8,17 +7,17 @@ import GlobalStateContext from "../Global/GlobalStateContext";
 const HomePage = () => {
   const { states, setters, requests } = useContext(GlobalStateContext);
 
-
   useEffect(() => {
     requests.getPokemons();
   }, [states, setters, requests]);
 
-  const addToPokedex = (newPokemon) => {
-    const index = states.pokemonsHome.findIndex((i) => i.id === newPokemon.id);
-    const newPokedex = [...states.pokedexList, newPokemon];
-    setters.setPokedexList(newPokedex);
-    states.pokemonsHome.splice(index, 1);
-  };
+
+  // const addToPokedex = (newPokemon) => {
+  //   const index = states.pokemonsHome.findIndex((i) => i.id === newPokemon.id);
+  //   const newPokedex = [...states.pokedexList, newPokemon];
+  //   setters.setPokedexList(newPokedex);
+  //   states.pokemonsHome.splice(index, 1);
+  // };
 
   return (
     <>
@@ -32,12 +31,12 @@ const HomePage = () => {
                 key={pokemon.id}
                 image={pokemon.sprites.front_default}
                 name={pokemon.name}
-                addToPokedex={() => addToPokedex(pokemon)}
+                pokemon={pokemon}
+                // addToPokedex={() => addToPokedex(pokemon)}
               />
           );
         })
       )}
-      
       </CardMain>
       
     </>
